requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
python-dotenv>=1.0.0
openpyxl>=3.1.0

